refactor: extract SystemMicrophonePermissionAccess.swift from MicrophonePermissionService.swift
Splits the `SystemMicrophonePermissionAccess` class (~140 lines, the
concrete MicrophonePermissionAccessing implementation) out of
MicrophonePermissionService.swift into its own file.

Byte-identical move of lines 6-146. Both files import AppKit + AVFAudio
+ AVFoundation + Foundation.

MicrophonePermissionService.swift: 326 → 185 lines. Tests: 667.
